[!!!][FEATURE] use rotating indices, so re-indexing can happen while the system is live
Behind the scenes, the index name which is configured is just an alias
to a timestamp-prefixed version of the index. On re-import, a new index
is created and when everything is complete, a switch towards the new index
is done.

This change is breaking because the "typo3cr" index needs to be deleted
manually when this change is merged.