[BUGFIX] Aliasing the index fails if name is used by actual index
If the index is created before an attempt is made to use an alias which
points to a suffixed alias name, the (re-)indexing of nodes failed when
the aliasing was attempted.

This change will check for any actual index being present and removes it
prior to creating an alias again.