[BUGFIX] After changing NodeType, ES indexes contains two documents
In ES the ID property is unique inside a single document type. With the
current implementation, after a NodeType change, we create a new document
with the new NodeType, but the old one stay in the index.

This change solve the issue with a DELETE query to check is an other
document share the same _id before indexing new values.