TASK: Implement BulkNodeIndexerInterface in NodeIndexer
This change implement the newly introduced BulkNodeIndexerInterface for
the TYPO3.TYPO3CR.Search package. When enabled this skip the check for
duplicate node with different node type. This feature is only useful when
changing the node type (live indexing). During the initial indexing this
can speedup the process quit a lots for big index.