TASK: Remove code that is never used
The case removed with this is never reached, since a removed node will
never be handled here. Instead, due to the way removal is handled in the
Content Repository and the way we use the signals emitted, a removed node
is either handled in removeNode() or not found for an update, since the
publication of a removal was already done at this point.