!!! TASK: Adjust the result of getSuggestions
Before the result of getSuggestions returned the suggestion of a search term if the term only consisted of one word. When the search term conisted of two words it returned the complete array. This is an absolutely unexpected behaviour. It was also only possible to acces the first defined suggest term.

This change now reurns the complete suggestion array, so it behaves like the comparable aggregation getter.