!!! TASK: Rename internal fields to neos prefixed
Previously fields that are not defined via properties have been started with an underscore or sometimes with two underscores. This is generally not a problem for elasticsearch but kibana treats these fields as elasticsearch-internal and doesn’t aloow to analyze them.

Now the fields follow the rules defined in https://www.elastic.co/guide/en/beats/devguide/current/event-conventions.html