BUGFIX: Set up missing alias during nodeindex:build
The command `nodeindex:build` does not set up aliases after creating
fresh indexes until it considers the indexing complete. This is good,
usually, since it allows for indexing "in the background". But if no
aliases exist at all, this is going to break.

This change fixes that by setting up aliases after creating (fresh)
indexes, if no aliases exist.

Fixes #370