FEATURE: Support `.simpleQueryStringFulltext()` simple query string query for fulltext search
We want to support phrase searches like `"Neos and Flow"`, where
quotes mean "I want this exact phrase" (similar to many search engines).

When using the .fulltext() query, this did not work, and we sometimes
encountered crashes when the query string contained quotes, f.e. `"Hallo"`
or `Ha"llo"`.

This is because the .fulltext() query does some magic for supporting UTF8
(as it seems), namely: `trim(json_encode($searchWord, JSON_UNESCAPED_UNICODE), '"')`.
This removes some quotes, but not all of them :-)

The simple_query_string query of Elasticsearch exposes a query string syntax
which the user is directly allowed to enter:

> While its syntax is more limited than the query_string query, the simple_query_string
> query does not return errors for invalid syntax. Instead, it ignores any invalid
> parts of the query string.

This is exactly what we want for a good search field behavior.

NOTE: We created an additional method because we were unsure about breaking
      backwards compatibility; but we think `simpleQueryStringFulltext()` should
      be used almost universally as replacement for `fulltext()`, as it returns
      better results and does not throw exceptions..